renamed base to base_camera
