+++ Camera/base_camera.py
